+++ src/TrueSheet.tsx
@@ -404,23 +404,11 @@ export class TrueSheet
       return Math.min(1, detent);
     });
 
-    const containerStyle =
-      this.props.scrollable &&
-      Platform.select({
-        android: styles.scrollableAndroidContainer,
-      });
-
-    const contentStyle =
-      this.props.scrollable &&
-      Platform.select({
-        android: styles.scrollableAndroidContent,
-      });
-
     return (
       <TrueSheetViewNativeComponent
         {...rest}
         ref={this.nativeRef}
-        style={styles.sheetView}
+        style={[StyleSheet.absoluteFill, styles.sheetView]}
         detents={resolvedDetents}
         backgroundBlur={backgroundBlur}
         blurOptions={blurOptions}
@@ -459,13 +447,15 @@ export class TrueSheet
         onBackPress={this.onBackPress}
       >
         {this.state.shouldRenderNativeView && (
-          <TrueSheetContainerViewNativeComponent style={containerStyle}>
+          <TrueSheetContainerViewNativeComponent style={scrollable && StyleSheet.absoluteFill}>
             {header && (
               <TrueSheetHeaderViewNativeComponent style={[styles.header, headerStyle]}>
                 {isValidElement(header) ? header : createElement(header)}
               </TrueSheetHeaderViewNativeComponent>
             )}
-            <TrueSheetContentViewNativeComponent style={[style, contentStyle]}>
+            <TrueSheetContentViewNativeComponent
+              style={[style, scrollable && styles.scrollableContent]}
+            >
               {children}
             </TrueSheetContentViewNativeComponent>
             {footer && (
@@ -484,22 +474,8 @@ const styles = StyleSheet.create({
   sheetView: {
     zIndex: -9999,
     pointerEvents: 'box-none',
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-
-    // Android needs a fixed bottom to avoid jumping content
-    bottom: Platform.select({ android: 0 }),
   },
-  scrollableAndroidContainer: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
-  },
-  scrollableAndroidContent: {
+  scrollableContent: {
     flexGrow: 1,
     flexBasis: 0,
   },